bike_rental/backend/requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
pony==0.7.17
psycopg2-binary==2.9.7
python-dotenv==1.0.0
python-dateutil==2.8.2